refactor(Flagsmith on Flagsmith): Write the environment document with the Flagsmith CLI
The updater fetched the environment document itself and masked it in
Python. The CLI does the fetching, and the masking is a jq expression in
the workflow, so the API no longer carries code that only ever ran in CI.

Authentication changes with it: the environment-document endpoint took a
server-side key, whereas the CLI resolves the environment over the Admin
API and needs a Master API key.

beep boop
