+++ .github/workflows/update-flagsmith-environment.yml
@@ -5,40 +5,63 @@ on:
     - cron: 0 8 * * *
   workflow_dispatch:
 
-defaults:
-  run:
-    working-directory: api
-
 permissions:
   contents: read
+
+env:
+  # TODO: change to 12 when the projects merge — https://github.com/Flagsmith/flagsmith/issues/7988
+  FLAGSMITH_PROJECT: "11737"
+  FLAGSMITH_DOCUMENT_ENVIRONMENT: Production
+  FLAGSMITH_DOCUMENT_PATH: api/integrations/flagsmith/data/environment.json
 
 jobs:
   update_server_defaults:
     runs-on: depot-ubuntu-latest
     name: Update API Flagsmith Defaults
     env:
-      FLAGSMITH_ON_FLAGSMITH_SERVER_API_URL: https://edge.api.flagsmith.com/api/v1
-      FLAGSMITH_ON_FLAGSMITH_SERVER_KEY: ${{ secrets.FLAGSMITH_ON_FLAGSMITH_SERVER_KEY }}
+      FLAGSMITH_API_KEY: ${{ secrets.FLAGSMITH_API_KEY }}
 
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@fbc6f3992d24b796d5a048ff273f7fcc4a7b6c09 # v5.1.0
 
-      - name: Install uv
-        uses: ./.github/actions/install-uv
+      - name: Generate GitHub App token
+        id: app-token
+        uses: actions/create-github-app-token@bcd2ba49218906704ab6c1aa796996da409d3eb1
         with:
-          working-directory: api
+          client-id: ${{ vars.FLAGSMITH_ENGINEERING_GH_APP_ID }}
+          private-key: ${{ secrets.FLAGSMITH_ENGINEERING_GH_APP_PRIVATE_KEY }}
 
-      - name: Install Dependencies
-        run: make install-packages
+      - name: Install Flagsmith CLI
+        run: curl -fsSL https://raw.githubusercontent.com/Flagsmith/flagsmith-cli/main/install.sh | sh
 
       - name: Update defaults
-        run: uv run python manage.py updateflagsmithenvironment
+        run: |
+          flagsmith environment document "$FLAGSMITH_DOCUMENT_ENVIRONMENT" --no-input --jq '{
+            id: 0,
+            api_key: "masked",
+            name,
+            use_identity_composite_key_for_hashing,
+            feature_states,
+            project: {
+              id: 0,
+              name: .project.name,
+              hide_disabled_flags: .project.hide_disabled_flags,
+              segments: [],
+              organisation: {
+                id: 0,
+                name: .project.organisation.name,
+                feature_analytics: .project.organisation.feature_analytics,
+                stop_serving_flags: .project.organisation.stop_serving_flags,
+                persist_trait_data: .project.organisation.persist_trait_data
+              }
+            }
+          }' | jq . > "$FLAGSMITH_DOCUMENT_PATH"
 
       - name: Create Pull Request
-        uses: peter-evans/create-pull-request@v8
+        uses: peter-evans/create-pull-request@5f6978faf089d4d20b00c7766989d076bb2fc7f1 # v8.1.1
         with:
           commit-message: Update API Flagsmith Defaults
           delete-branch: true
-          title: 'chore: Update Flagsmith environment document'
+          title: "chore: Update Flagsmith environment document"
           labels: api
-          token: ${{ secrets.UPDATE_ENVIRONMENT_DOCUMENT_WORKFLOW_GH_PAT }}
+          token: ${{ steps.app-token.outputs.token }}
