+++ api/integrations/flagsmith/constants.py
@@ -1,0 +1,6 @@
+import os
+
+ENVIRONMENT_JSON_PATH = os.path.join(
+    os.path.dirname(os.path.realpath(__file__)), "data/environment.json"
+)
+"""Offline defaults for Flagsmith on Flagsmith evaluation."""
